backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
supabase>=2.3.0
google-generativeai>=0.8.0
google-genai>=1.0.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.24.0
sse-starlette>=2.0.0
python-jose[cryptography]>=3.3.0
aiofiles>=23.2.1
mangum>=0.17.0
